pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notion-clerk"
version = "0.1.0"
description = "Chat with your Notion workspace in plain English"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "responses>=0.25.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests that require live API keys (deselect with '-m not integration')",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["_archive/", "*.ipynb"]
